extract DurationTimer
